Split delegate twins out of RadarLocationManager+Swift
Pure code motion, no behavior change. The visit port pushed
RadarLocationManager+Swift.swift to 417 lines, over SwiftLint's 400-line
file_length limit.

Moves the CLLocationManagerDelegate-callback twins (didUpdateHeading,
didChangeAuthorizationStatus, didFail, effectiveLocation, shouldHandleRegion,
shouldHandleVisit, locationSource) into
RadarLocationManager+SwiftDelegate.swift, leaving the tracking- and
region-management twins with the identifier prefix constants. 315 and 123
lines respectively.

identifierPrefix drops `private` because shouldHandleRegion now lives in the
other file and Swift's `private` is file-scoped. The other four prefix
constants stay private.
