+++ RadarSDK/RadarLocationManager+Swift.swift
@@ -25,7 +25,11 @@ final class RadarLocationManagerSwift: NSObject {
 
     // Mirror of the identifier prefix constants in RadarLocationManager.m. Kept in sync by
     // hand until that file is fully ported.
-    private static let identifierPrefix = "radar_"
+    //
+    // `identifierPrefix` is internal rather than private because `shouldHandleRegion` lives in
+    // RadarLocationManager+SwiftDelegate.swift, and Swift's `private` is file-scoped. The rest
+    // stay private — they're only read by the sync/remove methods below.
+    static let identifierPrefix = "radar_"
     private static let bubbleGeofenceIdentifierPrefix = "radar_bubble_"
     private static let syncGeofenceIdentifierPrefix = "radar_geofence_"
     private static let syncBeaconIdentifierPrefix = "radar_beacon_"
@@ -262,22 +266,6 @@ final class RadarLocationManagerSwift: NSObject {
         }
     }
 
-    // Takes the heading rather than the location manager, so it uses a plain selector
-    // instead of the `...OnLocationManager:` convention. `RadarState` is a non-@objc Swift
-    // class and can't appear in an @objc signature, so it's constructed inside the method.
-    @objc(didUpdateHeading:)
-    static func didUpdateHeading(_ heading: CLHeading) {
-        RadarState().lastHeadingData = [
-            "magneticHeading": heading.magneticHeading,
-            "trueHeading": heading.trueHeading,
-            "headingAccuracy": heading.headingAccuracy,
-            "x": heading.x,
-            "y": heading.y,
-            "z": heading.z,
-            "timestamp": heading.timestamp.timeIntervalSince1970,
-        ]
-    }
-
     @objc(shutDownOnLocationManager:lowPowerLocationManager:)
     static func shutDown(locationManager: CLLocationManager, lowPowerLocationManager: CLLocationManager) {
         RadarLogger.shared.debug("🦅 Shutting down")
@@ -291,32 +279,6 @@ final class RadarLocationManagerSwift: NSObject {
         RadarLogger.shared.debug("🦅 Requesting location")
 
         locationManager.requestLocation()
-    }
-
-    @objc(didChangeAuthorizationStatus:)
-    static func didChangeAuthorizationStatus(_ status: CLAuthorizationStatus) {
-        let state = RadarState()
-        let previousStatus = state.locationAuthorizationStatus
-        state.locationAuthorizationStatus = status
-
-        if status == previousStatus {
-            return
-        }
-
-        guard let config = RadarSettings.sdkConfiguration else {
-            return
-        }
-        guard status == .authorizedAlways || status == .authorizedWhenInUse,
-            config.trackOnceOnAppOpen || config.startTrackingOnInitialize
-        else {
-            return
-        }
-
-        RadarLogger.shared.log(level: .info, message: "🦅 Location services authorized")
-        Radar.trackOnce(completionHandler: nil)
-        if config.startTrackingOnInitialize, !RadarSettings.tracking {
-            Radar.startTracking(trackingOptions: RadarSettings.trackingOptions)
-        }
     }
 }
 
@@ -335,23 +297,6 @@ extension RadarLocationManagerSwift {
         }
     }
 
-    @objc(didFailWithError:)
-    static func didFail(error: NSError) {
-        RadarLogger.shared.debug("🦅 CLLocation manager error | error = \(error)")
-        RadarSwift.bridge?.didFail(status: .errorLocation)
-    }
-
-    // Falls back to the last known location when the manager's current location is stale or
-    // missing — used by the region delegate callbacks (didEnterRegion/didExitRegion/
-    // didDetermineState) before they hand off to `handleLocation:`.
-    @objc(effectiveLocationForLocationManager:)
-    static func effectiveLocation(for locationManager: CLLocationManager) -> CLLocation? {
-        if let location = locationManager.location, location.isValid {
-            return location
-        }
-        return RadarSwift.bridge?.lastLocation()
-    }
-
     // Applies the remote tracking options carried by a `/config` or `/track` response's `meta`.
     // `RadarMeta.h` is a project-visibility header, so `RadarMeta` can't appear in an @objc
     // signature — the caller unwraps `meta.trackingOptions` (a public type) and keeps the
@@ -366,52 +311,5 @@ extension RadarLocationManagerSwift {
             RadarSettings.remoteTrackingOptions = nil
             RadarLogger.shared.debug("🦅 Removed remote tracking options | trackingOptions = \(Radar.getTrackingOptions())")
         }
-    }
-
-    // The shared entry gate for the region delegate callbacks: a region is ours only if it
-    // carries our identifier prefix, and we only act on it while tracking. `didEnterRegion`
-    // and `didExitRegion` ran identical copies of this, differing only in the log wording, so
-    // `action` ("entry"/"exit") is threaded through purely to preserve those messages.
-    // `didDetermineState` is deliberately not a caller — it gates on the beacon prefixes and
-    // has no tracking check.
-    @objc(shouldHandleRegionWithIdentifier:action:)
-    static func shouldHandleRegion(identifier: String, action: String) -> Bool {
-        guard identifier.hasPrefix(identifierPrefix) else {
-            RadarLogger.shared.debug("🦅 Ignoring region \(action): wrong prefix")
-            return false
-        }
-
-        guard RadarSettings.tracking else {
-            RadarLogger.shared.debug("🦅 Ignoring region \(action): not tracking")
-            return false
-        }
-
-        return true
-    }
-
-    // The guard half of `didVisit:`. The caller has already established that the manager has
-    // a location, so it's passed in rather than re-checked here — it's only needed for the
-    // log line. The `handleLocation:` tail stays in ObjC, so the decision is split across
-    // this predicate and `locationSource(for:)` below.
-    @objc(shouldHandleVisit:location:)
-    static func shouldHandleVisit(_ visit: CLVisit, location: CLLocation) -> Bool {
-        RadarLogger.shared.debug(
-            "🦅 Visit detected | arrival = \(visit.arrivalDate); departure = \(visit.departureDate); horizontalAccuracy = \(visit.horizontalAccuracy); visit.coordinate = (\(visit.coordinate.latitude), \(visit.coordinate.longitude)); manager.location = \(location)"
-        )
-
-        guard RadarSettings.tracking else {
-            RadarLogger.shared.debug("🦅 Ignoring visit: not tracking")
-            return false
-        }
-
-        return true
-    }
-
-    // Core Location reports a visit that is still in progress with `departureDate` set to
-    // the distant future, so that sentinel — not a nil check — is what distinguishes an
-    // arrival from a departure.
-    @objc(locationSourceForVisit:)
-    static func locationSource(for visit: CLVisit) -> RadarLocationSource {
-        return visit.departureDate == Date.distantFuture ? .visitArrival : .visitDeparture
     }
 }
